Add String+PathMatrchingTests unit tests
